Maybe now?

+++ src/backend/libc/process/syscalls.rs
@@ -152,7 +152,7 @@ pub(crate) fn getpgid(pid: Option<Pid>) -> Pid {
     unsafe {
         let pgid = c::getpgid(Pid::as_raw(pid) as _);
         debug_assert_ne!(pgid, 0);
-        Pid::from_raw_nonzero(RawNonZeroPid::new_unchecked(pgid as u32))
+        Pid::from_raw_nonzero(RawNonZeroPid::new_unchecked(pgid))
     }
 }
